requirements.txt
fastapi~=0.115.7
aiologger[aiofiles]~=0.7.0
pip-chill
uvicorn
gunicorn
python-dotenv~=1.0.1
yandex-cloud-ml-sdk
beautifulsoup4~=4.12.3
pydantic~=2.10.6